NationBuilder signup form with extra field
This is a customized signup template that saves extra information to
the Background section of a NationBuilder profile. I’m using the html
input method with ‘signup[note]’. There’s some extra code that checks
if the signup is a member that can be removed, and I’ve also chosen to
hide the opt in checkboxes.

Example at http://www.teamacetennis.com/first_year_free